Changes the Dyna.Core integration test folder to match the name of the project.
