+++ tests/Dyna.Core.Tests.Integration/Dyna.Core.Tests.Integration.csproj
